add: create_objective sub-package
